Remove all subsystem unit test. Now rely on tests/unittest
