+++ src/Configuration/CFGCommon/CMakeLists.txt
@@ -21,7 +21,6 @@
 
 cmake_minimum_required(VERSION 3.15)
 set(subsystem cfgcommon)
-set(test_bin ${subsystem}_test)
 
 project(${subsystem} LANGUAGES CXX)
 
@@ -43,8 +42,6 @@ endif()
 # 1. auto generate all the argument file
 #
 # 2. subsystem library (in this case cfgcommon) - it contains all the core code to help configuration. It will be called by other elements
-#
-# 3. test_bin executable (in this case cfgcommon_test exe) - it is external command line entry which eventually call subsystem library (item 1) for unit testing
 #
 #################################################################################################################
 
@@ -70,17 +67,6 @@ add_library(
   CFGArg.cpp
   ${CMAKE_CURRENT_BINARY_DIR}/CFGArg_auto.h
 )
-
-###################
-#
-# test_bin which also has dependency on its own subsystem library
-#
-###################
-add_executable(
-  ${test_bin}
-  Test/CFGCommon_test.cpp
-)
-target_link_libraries(${test_bin} ${subsystem})
 
 ###################
 #
